Combine all summary related functionality in one file
